# src/main/java/au/com/totemsoft/b2/v1/model/ShipmentRequest.java
package au.com.totemsoft.b2.v1.model;

import java.util.Objects;

import com.fasterxml.jackson.annotation.JsonProperty;

import io.swagger.annotations.ApiModelProperty;

/**
 * ShipmentRequest
 */
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
public class ShipmentRequest   {
  @JsonProperty("id")
  private String id;

  public ShipmentRequest id(String id) {
    this.id = id;
    return this;
  }

  /**
   * Entity Id (primary key)
   * @return id
  */
  @ApiModelProperty(example = "S-5165614", value = "Entity Id (primary key)")


  public String getId() {
    return id;
  }

  public void setId(String id) {
    this.id = id;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    ShipmentRequest shipmentRequest = (ShipmentRequest) o;
    return Objects.equals(this.id, shipmentRequest.id);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class ShipmentRequest {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}